requirements.txt
insightface
opencv-python
faiss-cpu
numpy
sqlalchemy
python-dotenv
SpeechRecognition==3.10.0
PyAudio==0.2.13
mediapipe==0.10.9
opencv-python==4.8.1.78
numpy
